% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farey_bins.r
\name{farey_bins}
\alias{farey_bins}
\title{Create bins with a farey sequence}
\usage{
farey_bins(n)
}
\arguments{
\item{n}{A number.}
}
\value{
A list with the bin edges and bin centres for the resulting sequence.
}
\description{
Create bins with a farey sequence
}
