% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ideal_hist.r
\name{gen_ideal_hist}
\alias{gen_ideal_hist}
\title{Generate V(pr) model distributions}
\usage{
gen_ideal_hist(matrix, bins, layer_map)
}
\arguments{
\item{matrix}{A data frame with the following columns: CHR, POS, REF, ALT, Ttr, Ntr, Nex, Tex}

\item{bins}{A list outputted from \code{farey_bins()}}

\item{layer_map}{A named character vector with the column names corresponding to each layer (Nex, Ntr, Tex, Ttr).}
}
\value{
A list with the bootstrapped V(pr) model distributions and empirical cumulative density functions for each of the layers.
}
\description{
Generate V(pr) model distributions
}
