% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaf.r
\name{get_vaf}
\alias{get_vaf}
\title{Calculate VAF}
\usage{
get_vaf(ref, alt)
}
\arguments{
\item{ref}{A numeric vector with number of reference reads for a set of loci}

\item{alt}{A numeric vector with numer of variant reads for a set of loci}
}
\value{
The Variant Allele Frequency of a locus given reference and altered (variant) reads for a given locus
}
\description{
Calculate VAF
}
\examples{
alt <- sample(10:1000, 5)
ref <- sample(10:1000, 5)
vaf <- get_vaf(ref, alt)
}
